+++ src/node_values.rs
@@ -1,3 +1,3 @@
 use std::collections::HashMap;
 
-pub(crate) type NodeValues = HashMap<usize, f64>;
+pub(crate) type NodeValues = HashMap<usize, f32>;
