+++ src/nodes/output_node.rs
@@ -8,7 +8,7 @@ pub(crate) struct InputIds {
 /// Saves the input value in a accessible way
 #[derive(Debug)]
 pub(crate) struct OutputNode {
-    value: f64,
+    value: f32,
     pub(crate) id: usize,
     pub(crate) input_ids: InputIds,
 }
@@ -22,13 +22,13 @@ impl OutputNode {
         }
     }
 
-    pub(crate) fn get_value(&self) -> f64 {
+    pub(crate) fn get_value(&self) -> f32 {
         self.value
     }
 }
 
 impl NodeTrait for OutputNode {
-    fn process(&mut self, node_values: &NodeValues) -> f64 {
+    fn process(&mut self, node_values: &NodeValues) -> f32 {
         let input = node_values.get(&self.input_ids.input).unwrap();
         self.value = *input;
         self.value
