+++ src/nodes/subtract_node.rs
@@ -13,7 +13,7 @@ pub(crate) struct SubtractNode {
 }
 
 impl NodeTrait for SubtractNode {
-    fn process(&mut self, node_values: &NodeValues) -> f64 {
+    fn process(&mut self, node_values: &NodeValues) -> f32 {
         let input1 = node_values.get(&self.input_ids.input1).unwrap();
         let input2 = node_values.get(&self.input_ids.input2).unwrap();
         input1 - input2
